pyproject.toml
[tool.poetry]
name = "fed-sentiment"
version = "0.1.0"
description = ""
authors = ["benisbuzz <ben@danziger.uk>"]
readme = "README.md"
packages = [{include = "fed_sentiment"}]

[tool.poetry.dependencies]
python = ">=3.10, <3.13"
pandas = "^2.1.1"
openai = "^0.28.1"
python-dotenv = "^1.0.0"
ipython = "^8.16.0"
ipykernel = "^6.25.2"
structlog = "^23.1.0"
aiolimiter = "^1.1.0"
pathlib = "^1.0.1"
openpyxl = "^3.1.2"
certifi = "^2023.7.22"
mypy = "^1.5.1"
pandas-stubs = "^2.1.1.230928"
tiktoken = "^0.5.1"
tqdm = "^4.66.1"
pydantic = "^2.4.2"
pypdf2 = "^3.0.1"
transformers = "^4.34.0"
nltk = "^3.8.1"
matplotlib = "^3.8.0"
fredpy = "^3.2.8"
statsmodels = "^0.14.0"
chardet = "^5.2.0"
fredapi = "^0.5.1"
fred = "^3.1"
beautifulsoup4 = "^4.12.2"
lxml = "^4.9.3"
ta-lib = "^0.4.28"
pyarrow = "^13.0.0"
numpy = "^1.26.0"
scikit-learn = "^1.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
